+++ src/cache/bing.ts
@@ -1,10 +1,10 @@
 import { ICacheManager } from "./interfaces/ICacheManager";
 import { BaseCache } from "./base";
-import { ILocationDictionary, IGeocoder, IGeocodeCoordinate, IGeocoderOptions } from "../geocoder/interfaces/geocoderInterfaces";
+import { ILocationDictionary, IGeocoder } from "../geocoder/interfaces/geocoderInterfaces";
 import { createGeocoder } from "../geocoder/geocoder";
 import { ILocationKeyDictionary } from "../interfaces/dataInterfaces";
 
-export class BingCache extends BaseCache implements ICacheManager {
+export class Bing extends BaseCache implements ICacheManager {
 
     private geocoder: IGeocoder;
 
@@ -16,15 +16,10 @@ export class BingCache extends BaseCache implements ICacheManager {
     public async loadCoordinates(keys: string[]): Promise<ILocationDictionary> {
         //const keys: string[] = Object.keys(keyDictionary);
         if (!keys || !keys.length) {
-            return new Promise<ILocationDictionary>((resolve, reject) => reject());
+            return new Promise<ILocationDictionary>((resolve, reject) => reject("Empty location keys"));
         }
 
         return this.geocoder.geocodeByDataFlow(keys);
-    }
-
-    // rewrite details
-    private async loadCoordinateFromBing(key: string, category?: string): Promise<IGeocodeCoordinate> {
-        return this.geocoder.geocode({ query: key, category: category } as IGeocoderOptions);
     }
 
     public async saveCoordinates(coordinates: ILocationDictionary): Promise<string> {