+++ src/cache/utils/utils.ts
@@ -1,6 +1,0 @@
-export function getShortKey(key: string): string {
-    if (!key || !key.length) {
-        return key;
-    }
-    return key.match(/([^;]+)$/i)[0].replace(/\/$/g, '').trim();
-}