+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * UPD: using StorageService for localStorage logic
 * UPD: powerbi-visuals-tools 3.0.9 for StorageService usage
 * UPD: powerbi-visuals-utils-testutils 2.1.6 for StorageService mocks usage
+* Refactoring of coordinates loading: from memory, StorageService and Bing Geocode DataFlow API
 
 ## 2.5.1
 * FIX: shifting for camera position was repaired